refactor: remove form-layout demo module
