+++ vaadin-form-layout-flow-parent/vaadin-form-layout-flow-integration-tests/src/main/java/com/vaadin/flow/component/formlayout/test/FormLayoutView.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.formlayout.demo;
+package com.vaadin.flow.component.formlayout.test;
 
 import java.io.Serializable;
 import java.time.LocalDate;
@@ -28,9 +28,11 @@ import com.vaadin.flow.component.formlayout.FormLayout;
 import com.vaadin.flow.component.formlayout.FormLayout.FormItem;
 import com.vaadin.flow.component.formlayout.FormLayout.ResponsiveStep;
 import com.vaadin.flow.component.html.Div;
+import com.vaadin.flow.component.html.H2;
 import com.vaadin.flow.component.html.Label;
 import com.vaadin.flow.component.html.NativeButton;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
+import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.component.textfield.TextField;
 import com.vaadin.flow.data.binder.Binder;
 import com.vaadin.flow.data.binder.Binder.Binding;
@@ -39,7 +41,6 @@ import com.vaadin.flow.data.binder.BindingValidationStatus;
 import com.vaadin.flow.data.validator.EmailValidator;
 import com.vaadin.flow.data.validator.StringLengthValidator;
 import com.vaadin.flow.data.value.ValueChangeMode;
-import com.vaadin.flow.demo.DemoView;
 import com.vaadin.flow.function.SerializablePredicate;
 import com.vaadin.flow.router.Route;
 
@@ -49,10 +50,9 @@ import com.vaadin.flow.router.Route;
  * @author Vaadin Ltd
  */
 @Route("vaadin-form-layout")
-public class FormLayoutView extends DemoView {
+public class FormLayoutView extends Div {
 
-    @Override
-    public void initView() {
+    public FormLayoutView() {
         createResponsiveLayout();
         createFormLayoutWithItems();
         createLayoutHandleColspans();
@@ -372,4 +372,11 @@ public class FormLayoutView extends DemoView {
         addCard("Using form layout inside a composite", layout);
     }
 
+    private void addCard(String title, Component... components) {
+        VerticalLayout layout = new VerticalLayout();
+        layout.setMargin(true);
+        layout.add(new H2(title));
+        layout.add(components);
+        add(layout);
+    }
 }
