+++ vaadin-form-layout-flow-parent/vaadin-form-layout-flow-integration-tests/src/test/java/com/vaadin/flow/component/formlayout/tests/FormLayoutIT.java
@@ -23,17 +23,23 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.Dimension;
 import org.openqa.selenium.WebElement;
 
-import com.vaadin.flow.component.formlayout.demo.FormLayoutView;
-import com.vaadin.tests.ComponentDemoTest;
+import com.vaadin.flow.testutil.TestPath;
+import com.vaadin.testbench.TestBenchTestCase;
+import com.vaadin.tests.AbstractComponentIT;
 
 /**
  * Integration tests for the {@link FormLayoutView}.
  */
-public class FormLayoutIT extends ComponentDemoTest {
+@TestPath("vaadin-form-layout")
+public class FormLayoutIT extends AbstractComponentIT {
+
+    private TestBenchTestCase layout;
 
     @Before
     public void init() {
+        open();
         Assert.assertTrue(isElementPresent(By.tagName("vaadin-form-layout")));
+        layout = this;
     }
 
     @Test
@@ -154,17 +160,7 @@ public class FormLayoutIT extends ComponentDemoTest {
                 executeScript("return arguments[0].checked;", element)));
     }
 
-    private void clearInput(String id) {
-        WebElement element = findElement(By.id(id));
-        executeScript("arguments[0].value = '';", element);
-    }
-
     private void forceClick(WebElement element) {
         executeScript("arguments[0].click();", element);
-    }
-
-    @Override
-    protected String getTestPath() {
-        return ("/vaadin-form-layout");
     }
 }
